Switch to rich-message module. Closes #65
