+++ lib/elements/messages.js
@@ -4,6 +4,19 @@ var h = require('virtual-dom/h')
 var inherits = require('util').inherits
 var BaseElement = require('./base-element')
 
+var htmlToVDom = require('html-to-vdom')
+var VNode = require('virtual-dom/vnode/vnode')
+var VText = require('virtual-dom/vnode/vtext')
+
+var convertHTML = htmlToVDom({
+  VNode: VNode,
+  VText: VText
+})
+
+function makeVDom (html) {
+  return convertHTML('<div class="text">' + html + '</div>')
+}
+
 function Messages (target) {
   BaseElement.call(this, target)
   this.shouldAutoScroll = true
@@ -28,7 +41,7 @@ Messages.prototype.render = function (channel, users) {
         h('.username', msg.username),
         verified ? h('.verified') : null,
         h('.timestamp', msg.timeago),
-        msg.vdom || h('.text', msg.text)
+        makeVDom(msg.html) || h('.text', msg.text)
       ])
     })
     childViews = h('.messages', {
